requirements.txt
astropy==5.2.2
geopandas==0.13.2
huggingface_hub[cli]==0.20.3
laspy[lazrs]==2.5.3
numpy==1.24.0
onnxruntime==1.17.0
opencv-python-headless==4.5.1.48
pyarrow==15.0.0
pyproj==3.5.0
python-dateutil==2.8.2
PyTurboJPEG==1.6.5
PyYAML==6.0.1
rasterio==1.2.10
scipy==1.10.1